메인 Wavey섹션의 반응형 버그 해결 & 컨텍트섹션 리팩토링
