+++ src/components/sections/Main.jsx
@@ -1,10 +1,10 @@
-import Wavey from "../Wavey";
+import WaveyContainer from "../../containers/WaveyContainer";
 import AboutMeContainer from "../../containers/AboutMeContainer";
 import Contact from "../Contact";
 function Main() {
   return (
     <main>
-      <Wavey />
+      <WaveyContainer />
       <AboutMeContainer />
       <Contact />
     </main>
