+++ src/redux/modules/index.js
@@ -1,7 +1,8 @@
 import { combineReducers } from "redux";
 import menu from "./menu";
 import about from "./about";
+import wavey from "./wavey";
 
-const reducer = combineReducers({ menu, about });
+const reducer = combineReducers({ menu, about, wavey });
 
 export default reducer;
