+++ src/styles/Contact.scss
@@ -22,179 +22,5 @@
     justify-content: center;
     align-items: center;
     margin-bottom: 5rem;
-
-    .business-card {
-      $width: 26rem;
-      min-width: 15rem;
-      $ratio: 55 / 85;
-
-      position: relative;
-      width: $width;
-      height: $width * $ratio;
-      border-radius: 0.5rem;
-      color: $reverse-font-color;
-      display: flex;
-      justify-content: center;
-      cursor: grab;
-
-      &.flipped {
-        .front {
-          transform: rotateY(180deg);
-        }
-        .back {
-          transform: rotateY(0deg);
-        }
-      }
-
-      .front,
-      .back {
-        position: absolute;
-        width: 100%;
-        height: 100%;
-        backface-visibility: hidden;
-        perspective: 600px;
-        background-color: #302f34;
-        padding: 2rem;
-        transition: 1s;
-        box-shadow: 0px -6px 8px 0px rgba(#000, 0.1),
-          0px 6px 8px 0px rgba(#000, 0.1), 6px 0px 8px 0px rgba(#000, 0.1),
-          -6px 0px 8px 0px rgba(#000, 0.1);
-      }
-
-      .front {
-        transform: rotateY(0deg);
-
-        .layer {
-          width: 100%;
-          height: 100%;
-          position: relative;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-
-          h1 {
-            font-size: 2rem;
-            font-weight: bold;
-            letter-spacing: -0.1rem;
-          }
-
-          .corner {
-            position: absolute;
-            backface-visibility: hidden;
-            width: 0.8rem;
-            height: 0.8rem;
-            border-right: 0.3rem solid $reverse-font-color;
-            border-top: 0.3rem solid $reverse-font-color;
-            border-radius: 0.3rem;
-
-            &:nth-of-type(1) {
-              right: 0;
-              top: 0;
-            }
-
-            &:nth-of-type(2) {
-              left: 0;
-              top: 0;
-              transform: rotateZ(-90deg);
-            }
-
-            &:nth-of-type(3) {
-              bottom: 0;
-              left: 0;
-              transform: rotateZ(180deg);
-            }
-
-            &:nth-of-type(4) {
-              bottom: 0;
-              right: 0;
-              transform: rotateZ(90deg);
-            }
-          }
-        }
-      }
-
-      .back {
-        transform: rotateY(-180deg);
-
-        .title {
-          padding-bottom: 4rem;
-
-          h2 {
-            font-size: $primary-sub-title-font-size;
-          }
-        }
-
-        .contents {
-          display: flex;
-          flex-direction: column;
-          align-items: baseline;
-          position: relative;
-
-          .content {
-            padding-bottom: 1rem;
-
-            .sub-title {
-              font-size: $primary-sub-title-font-size - 0.2rem;
-              color: #908f92;
-            }
-
-            .sub-content {
-              font-size: $primary-content-font-size;
-
-              a {
-                color: $reverse-font-color;
-              }
-            }
-
-            .email {
-              cursor: copy;
-            }
-
-            .tooltip {
-              visibility: hidden;
-              width: 3rem;
-              height: 1rem;
-              position: absolute;
-              top: 0;
-              left: 25%;
-              margin: auto;
-              padding: 0.2rem;
-              border-radius: 1rem;
-              font-size: 0.5rem;
-              text-align: center;
-              background-color: #{$reverse-background-color};
-
-              &::after {
-                content: "";
-                position: absolute;
-                top: 100%;
-                left: 50%;
-                margin-left: -0.5rem;
-                border-width: 0.3rem;
-                border-style: solid;
-                border-color: #{$reverse-background-color} transparent transparent
-                  transparent;
-              }
-            }
-          }
-        }
-
-        > .qr-code {
-          height: 100%;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          position: absolute;
-          right: 3rem;
-          top: 0;
-          bottom: 0;
-          margin: auto 0;
-
-          @media (max-width: $screen-size-mobile) {
-            display: none;
-          }
-        }
-      }
-    }
   }
 }
