+++ src/styles/Title.scss
@@ -15,7 +15,7 @@
 
     > * {
       margin: 0;
-      letter-spacing: 0.05rem;
+      letter-spacing: -0.05rem;
     }
 
     :first-child {
